Changed the Gemini PDA keyboard so that if key presses are processed first.
For example, Control + L_Shift + M & N can not be differentiated, so M will be blocked
until either Control or Shift is released.  The moment shift is released, we can
differentiate between M & N, but ensure that the press event comes first.  If M is
released first, the keystroke is lost due to the hardware design.